1033811: Added one class and proved code sample.
